Separate CommandBox testing for autocomplete as well as iteration of commands
